Stop linking in sqlite/_sqlite on Python 3
cscvs hasn't yet been ported to Python 3, and requires the old sqlite
module.  This may take some time to fix.  To avoid this blocking other
Python 3 work in the meantime, only link in the sqlite and _sqlite
system-installed packages on Python 2.

Rather than inventing a whole new syntax for this or splitting up the
system-packages.txt file, the simplest approach seemed to be to use the
PEP 508 environment marker syntax.  This also involved changing how
optional packages are specified.