Ignore 404 when deleting unwanted Swift files
librarian-gc has been consistently failing for some time with the
following pattern:

  2020-08-19 03:22:44 DEBUG   http://10.34.0.139:8080 "GET /v1/AUTH_xxx/librarian_6539?format=json&marker=3269994338 HTTP/1.1" 200 552359
  2020-08-19 03:22:44 DEBUG   http://10.34.0.139:8080 "GET /v1/AUTH_xxx/librarian_6539?format=json&marker=3269999999 HTTP/1.1" 200 2
  2020-08-19 03:22:45 DEBUG   http://10.34.0.139:8080 "DELETE /v1/AUTH_xxx/librarian_6539/3269500785 HTTP/1.1" 401 131
  2020-08-19 03:22:46 DEBUG   http://10.34.0.136:5000 "POST /v2.0/tokens HTTP/1.1" 200 3410
  2020-08-19 03:22:46 DEBUG   http://10.34.0.139:8080 "DELETE /v1/AUTH_xxx/librarian_6539/3269500785 HTTP/1.1" 404 70

The DELETE may happen in a different swiftclient.client.Connection
instance from the GETs, and the sequence of GETs to enumerate all the
files in Swift up to that point takes almost two hours on production at
the moment which is longer than Keystone's default token expiry time of
one hour, so it makes sense that the first DELETE might return 401.
However, it's not at all clear why the retried DELETE of the same object
gets 404 (nothing else in librarian-gc seems to have deleted that
object).  Although I'm concerned about this odd pattern, I've been
unable to reproduce this with the same version of Swift, so perhaps it's
best to just ignore it (since DELETE returning 404 still at least
indicates that the object is no longer present) and allow librarian-gc
to run to completion.