Export OCIRecipe.newPushRule(credentials_owner) on the API
It normally makes sense to allow controlling the owner of newly-created
objects such as OCI registry credentials in webservice API methods.
It's also more likely for the recipe owner to be a reasonable default
for the owner of new credentials than the logged-in user: team-owned
recipes will normally want team-owned credentials too.