Don't rely on Archive.signing_key
Archive.signing_key is specifically the GPGKey for an archive, and to
some extent is for backward compatibility purposes since the database
storage is on the signing_key_owner and signing_key_fingerprint columns
nowadays.  It makes less sense with keys stored on the signing service,
so avoid using it where we can.