Start fixing up email handling for Python 3
This is likely to be incomplete, as the email package has changed a lot
in Python 3; but this still appears to work in Python 2, and it cleans
up some things that impede getting the test suite working at all in
Python 3.

The main effect is to start trying to explicitly treat email messages as
bytes.