Use io.StringIO in SourcePackageRelease
This was only used in SourcePackageRelease.change_summary, and the input
is from a Unicode database column.

`io.StringIO(None).getvalue()` returns an empty string, while
`StringIO.StringIO(None).getvalue()` returns 'None'.  As a result, we
now display nothing rather than the string 'None' for SPRs with no
changelog_entry.  Although the rendering of this case as 'None' was
enforced by one test (written two years after the underlying code), I
think it was essentially accidental and that displaying nothing is
better.