Port unicode() calls in lp.soyuz to Python 3
This unfortunately requires some contextual clues, because
`six.text_type(b'foo')` returns `u'foo'` on Python 2 but `"b'foo'"` on
Python 3, while `six.ensure_text` works on bytes or text but not on
other types.  Use single-argument `six.text_type` in cases where we know
that the argument is not bytes, and `six.ensure_text` where we know the
argument is either bytes or text.