Drop GeoIP/GeoLite Legacy support
Production has been using GeoIP2 data provided by MaxMind via IS for a
while, so this was only for convenience on systems that don't have
access to that data.  I considered packaging the GeoLite2 Country
database in ppa:launchpad/ubuntu/ppa, but it comes with the licence
condition that redistributors must keep it up to date, which is too
onerous for a test package.  Instead, allow running with no GeoIP
database at all, and set the testrunner configuration to use a test
database derived only from the IP addresses used in the Launchpad test
suite and public WHOIS data.

The build script for the test database is in Perl, since I couldn't find
a Python writer library for the MaxMind database format; but I've also
committed the resulting database, since it's tiny.