Accept code import log file data over XML-RPC
The code import worker currently uploads its log file directly to the
librarian.  This is a bit of an obstacle to splitting out the code
import worker from the main Launchpad codebase: it wouldn't necessarily
be impossible to retain the remote librarian upload client in a future
lp-codeimport, but it would be somewhat awkward as that client is
currently designed with the expectation of having direct Launchpad
database access.

Log files aren't typically all that large, though, so it seems workable
to just send them as part of the XML-RPC `finishJobID` request and have
the webapp side of that request do the librarian upload instead.  This
prepares for that by accepting either XML-RPC-encapsulated binary data
or a librarian URL as the third parameter to `finishJobID`.