Run schema-management programs directly, not with bin/py
Running Python programs in the Launchpad tree with bin/py doesn't quite
work on Python 3; it's necessary to pass the -S flag as well to avoid
confusion with the wrong site module.  But it's also unnecessary
complexity, because all of these programs are executable and import
_pythonpath which will re-exec them with the correct Python executable
from the virtualenv if necessary; so just run them directly.