Work around readline crash in Python 3 on 18.04
On 18.04, importing GTK (which we do in lp.testing.html5browser)
indirectly loads libedit, and versions of Python 3 before 3.8.1 crash
when importing readline in a process that has libedit loaded
(https://bugs.python.org/issue38634,
https://bugs.launchpad.net/bugs/1899076).  As a result the test suite
crashes quite early on.

There's unfortunately no reasonable way to avoid this, because we aren't
loading libedit directly and nor are we importing readline directly;
they're both quite deeply-buried side-effects of other things.  We have
to take an unreasonable way instead: stub out the readline module in the
test suite on affected systems.  This is deliberately kept as narrow as
possible, and we should drop it if and when the Python bug is fixed in
Ubuntu 18.04 or we no longer care about trying to get Launchpad to run
on Python 3 on 18.04.