Rephrase excluded_scenarios for Python 3 compatibility
Per the Python Language Reference, "the scope of names defined in a
class block is limited to the class block; it does not extend to the
code blocks of methods".  In Python 3, list comprehensions are
implemented using a function scope, so they do not have access to names
defined in the containing class block.