Build constraints file dynamically
The constraints file used to specify Launchpad's Python dependencies is
logically built by combining a particular version of the Zope Toolkit's
ztk-versions.cfg file (which is published as a buildout configuration
file) with our own requirements, which may either override the ZTK
version of a particular package or add additional packages.  Since we
ported our build system from buildout to pip, we've been maintaining
this as a single file in PEP 508 requirements syntax with overrides
expressed by commenting out the original and adding another version
immediately afterwards; this was functional, but cumbersome and
confusing.

Instead, keep an unmodified copy of the relevant version of
ztk-versions.cfg, and apply local overrides to it dynamically.  This
should be much easier to maintain: we can upgrade to newer ZTK releases
just by dropping in a new version of ztk-versions.cfg (and perhaps
removing overrides that are no longer necessary), and we can upgrade our
local requirements without needing to worry about whether doing so
overrides ZTK-specified versions.

When building on Python 2, there is no functional change.

When building on Python 3, Pygments, Sphinx, and
sphinxcontrib-websupport get upgraded versions.  This was already
specified in ztk-versions.cfg, but previously went unnoticed in the
manual conversion to PEP 508 requirements syntax.