Don't materialise all archives in sync-signingkeys
I added this list() call due to an inchoate thought that it might be
needed to do stable iteration over the list of archives, but in fact
there's no good reason for that since we're ordering by Archive.id
anyway.  Loading all the archives into the Storm cache up-front made
commits unreasonably slow.