number-cruncher: Count builders as building if they have a build queue
We don't really need to check the build queue status to determine that a
builder is building; just having a build queue is sufficient, even if
it's e.g. in the process of a cancelling a build.  Compare the logic in
SlaveScanner.scan.

This also fixes occasional LostObjectError exceptions when a build
completes between the start of the transaction taken by
`self.builder_factory.update()` and the start of the transaction taken
by `self._updateBuilderCounts()`, since evaluating
`builder.build_queue.status` can't be done solely using the prefetched
builder vitals and requires a database query.