Add configurable distribution default traversal rules
We can use redirect_default_traversal to configure whether the canonical
URL for a distroseries is (e.g.) /ubuntu/focal or /ubuntu/+series/focal.
In the future we'll also be able to use default_traversal_policy to
configure the default traversal for a distribution to mean something
else: for example it might well make sense to use it for distribution
source packages rather than for series.