Fix get_email_notifications tests to be effective
The tests for get_email_notifications didn't test what they claimed to
test, because refactorings of the code under test some years ago meant
that getBugNotificationRecipients is no longer called there, and it
looks as though the tests were fixed up incorrectly.  Fix the tests to
actually be effective, and add assertions on the log output to ensure
that this can't regress in the same way in future.