Make test_gen_reload_queries_with_mixed_stores use a list
Objects corresponding to the same database row from different stores
compare equal according to SQLBase.__eq__.  They currently don't *hash*
equal, but we'll be fixing that soon.  Once we do, trying to keep track
of them in a set will produce incorrect results for the purposes of this
test, so use a list instead.