Define SQLBase.__hash__
On Python 2, SQLBase objects were hashable without this, but they broke
the contract for __hash__, because objects from different stores with
the same class and id would compare equal but have different hash
values.

Python 3 made it more difficult to make this kind of mistake, by
requiring that objects that define __eq__ must also define __hash__ in
order to be hashable, so now we do so.  However, the case of
newly-created objects where we haven't yet got a sequence value back
from the database requires some care.  We can no longer work around this
using object identity as was done by SQLBase.__eq__, since that would
lead to the object's hash value changing once its id is known.  Instead,
we flush the store if we don't yet know the object's id; this is perhaps
slightly surprising in a special method, but it seems the least bad
approach.

To minimise confusion, also rework SQLBase.__eq__ to use the same
flush-the-store-if-necessary approach rather than checking object
identity.