Remove excessive decoding from FileUploadProtocol
Broken by 2d3fa6f225bea7b4c2393810454cca10439027b7, because lineReceived
now decodes the line first so we shouldn't decode it again later.  We
get away with this on Python 2 because (confusingly) it's usually
possible to decode a unicode object there as long as it only contains
ASCII, but Python 3 is stricter.