Allow __len__ on several model classes
The reasoning is as in c1c601778ff02dd26e127551d2a4dd047affef5c: on
Python 3, tuple() and list() call len() internally to preallocate the
correct amount of space if possible.  If __len__ is unimplemented, then
they fall back to guessing, but ForbiddenAttribute confuses them and
results in failing to create the list.  Allowing __len__ through the
security proxy, even though it's unimplemented, avoids this.