Make BuilderSet.preloadProcessors more robust
We occasionally see buildd-manager log entries like this:

  Failure while updating builders:
  Traceback (most recent call last):
    File ".../lib/lp/buildmaster/manager.py", line 755, in scanBuilders
      self.builder_factory.update()
    File ".../lib/lp/buildmaster/manager.py", line 274, in update
      [b for b, _ in builders_and_current_bqs])
    File ".../lib/lp/buildmaster/model/builder.py", line 287, in preloadProcessors
      cache._processors_cache.append(store.get(Processor, pid))
  AttributeError: 'DefaultPropertyCache' object has no attribute '_processors_cache'

The oldest occurrence of this on dogfood was 2020-10-01, and the oldest
occurrence on production was on 2020-10-10 (though there may have been
some older occurrences on production that have expired from logs).  The
frequency is quite low: 11 occurrences on production including the first
on 2020-10-10.

It seems theoretically possible for the Storm cache to return a
different object from `store.get(Builder, bid)` than the Builder object
with the same ID that was passed into `BuilderSet.preloadProcessors`,
although it's not quite clear exactly how and I haven't been able to
reproduce it in the test suite.  However, it's straightforward enough to
guarantee to update the processors cache on the objects that were passed
in regardless of whatever Storm cache shenanigans may occur, so do so.