Remove cscvs
Now that lp.codehosting.codeimport has been split out to a separate
lp-codeimport project, the only remaining use of cscvs in the main
Launchpad tree was to validate CVSROOTs.  This isn't worth the fairly
significant build system complexity it incurs, and it's an obstacle to
running on Python 3, so just inline a cleaned-up version of the
relatively small amount of parsing code involved instead.

LP: #1016334