Work around Python 3's round() semantics
Python 3 changed the built-in round() function to use round-to-even for
exact half-way cases.  While this eliminates a bias away from zero and
is thus a better choice in many cases, it's not what several parts of
Launchpad expected.  Use the decimal module instead, which produces
identical results for our purposes on both Python 2 and 3.