publish-distro: Reset Storm store between each archive
When publishing large archives, a large number of alive objects can be
accumulated in the Storm store, which means that each commit can take a
long time as it has to iterate over every alive object to invalidate it.
This has been particularly noticeable on production since raising the
Storm cache size for LPCONFIG=ppa-publish to 500000 from the default
value of 10000, which has fixed preloading at the expense of slowing
down commits: after publishing some large archives, each commit can take
on the order of ten seconds, and since there are typically four commits
per dirty PPA this can add up very quickly.

To avoid this, reset the store between archives in roughly the same way
we do between web requests.  This means we have to go to some more work
in tests to reload objects from the database.

LP: #1910422