Fix setting of launchpad.publicationthreadduration on Python 3
We forgot to set it in LaunchpadBrowserPublication.afterCall.

While we're here, rearrange the other settings of
launchpad.{traversal,publication}threadduration so that they immediately
follow the corresponding settings of
launchpad.{traversal,publication}duration rather than having a statsd
call in between, to make it (at least in my mind) a little easier to
follow.