Ensure that private teams are visible on +sharing
Product:+sharing and Distribution:+sharing are only visible to pillar
drivers, who should be able to see who has access to their pillar.
However, it's possible to end up with a pillar having a grant to a
private team that not all the drivers can see.  In some ways it might be
best to grant LimitedView in this case, but doing that in the security
adapter would be risky for performance as that adapter is already very
complex and often slow.  Instead, precache LimitedView permissions for
all grantees just for PillarSharingView.

LP: #1913334