Remove bogus test for BugSubscriptionFilter without subscription
BugSubscriptionFilter.structural_subscription is marked as
allow_none=False, so it doesn't really make sense to test its behaviour
without a related subscription.  In fact there are no rows where this
column is NULL on any of our instances, so we ought to make the column
formally NOT NULL in the database.  Removing this test is the first step
towards being able to do that.