Fix lp.registry.model.projectgroup doctests
Broken by converting to IGNORE_EXCEPTION_MODULE_IN_PYTHON2, mainly
because I forgot to tell it to use the correct checker.

NORMALIZE_WHITESPACE is required because OutputChecker seems to get a
bit confused by an extra newline at the end of examples found in
docstring-based doctests.  This is probably a bug in zope.testing, but
since it's only needed for migration from Python 2 and is easy enough to
work around like this, it doesn't seem worth the effort to fix properly.