Rework authentication tests for bug command emails
TestSignatureTimestampValidation.test_bad_timestamp_but_no_commands
didn't work on Python 3 because it's no longer possible to clobber the
SignedMessage.signature property.  However, on inspection, this test was
only passing by accident anyway, and substituting an object with a
timestamp for SignedMessage.signature couldn't have worked, because
SignedMessage.signature is supposed to be a byte string.  It looks as
though the author of this test confused SignedMessage.signature with the
return value of GPGHandler.getVerifiedSignature.  Furthermore, the test
didn't actually check the signature timestamp, because it calls the
handler directly rather than first calling authenticateEmail; and a
timestamp too far in the future in fact results in an InvalidSignature
exception from authenticateEmail rather than successfully processing
command-free emails.

Rework this whole test suite so that it tests something meaningful,
namely that strong authentication is required to process commands in bug
emails.  We now also include a test for a scenario that involves
rejecting an insufficiently-authenticated email, making it easier to
notice when tests don't really test what they claim to test.