Strip PGP signature from .changes files in more places
We've done this for a long time in order to avoid replay attacks where
people extract a .changes file and use it to upload the same thing to a
different archive, but notifications for acceptances that didn't require
manual approval weren't appropriately stripped.  Fix
PackageUpload.notify to remove the signature.

LP: #1915002