Tighten up IGitRef __eq__, __ne__, and __hash__
As part of Python 3 support, we need to implement proper __eq__, __ne__,
and __hash__ methods for StormBase.  Before we can do that, we first
need to tighten up the custom implementations for GitRef and
GitRefRemote, which need their own specialized rules.

The main functional changes from this commit alone are that
GitRefDefault and GitRefFrozen instances can be compared with GitRef
instances and hashed, and that GitRefRemote instances now have correct
hashes derived only from their repository URL and path.