Reduce dependencies on flattened email message representation
There are various minor differences in the flattened email message
representation produced by `message.as_string()` and similar methods
between Python 2 and 3: for instance, quoted-printable encoding of
whitespace at the end of a line is handled slightly differently.  In
most cases we only really care that the message decodes to the expected
content, so adjust various tests to compare the decoded output.