Fix handling of removed objects in StormBase
Objects that have been removed don't have `primary_vars`, and `Store.of`
returns None for them.  This caused `StormBase.__eq__` (and, less
importantly, `StormBase.__hash__`) to crash if given such an object.  In
most cases we don't keep removed objects around, but there is at least
one exception: the cache of the current `BuildQueue` associated with the
builder managed by a `SlaveScanner`, which is checked for equality with
the `BuildQueue` found from the builder's vitals.

To avoid this problem, use `IStore(obj.__class__).flush()` instead, and
in the `__eq__` case explicitly return False if this still doesn't cause
`primary_vars` to appear.  (`__hash__` will still crash if used on a
removed object, but that seems to be OK.)  This is a better match for
the details of the behaviour of `SQLBase`.

Fixes a regression introduced by
6f2a36c1e74d99f511b033c19273c4cb2b67c6c4.

LP: #1916522