Minimally fix BugTaskSearchBugsElsewhereTest for Python 3
BugTaskSearchBugsElsewhereTest is only used from
lib/lp/bugs/stories/bugs/xx-bugs-advanced-search-upstream-status.txt.
Using a unittest.TestCase from a doctest is pretty dodgy, and on Python
3 the assertEqual method doesn't work.

The ideal fix would be to rewrite the doctest as a unittest, merging
this helper class into it along the way.  However, for now, avoiding
TestCase methods gets the test working.