Avoid construction order problems with Distribution
On Python 3.5, dict ordering is randomized and the order of keyword
arguments is not preserved (the latter is fixed in Python 3.6), so
SQLBase.__init__ may set arguments in any order, which can cause
problems if e.g. owner validation fails before display_name is set and
thus __repr__ fails.  We'll want an explicit constructor when we convert
Distribution to StormBase anyway, so do that now to fix the order in
which attributes are set and handle errors properly.

This is made a little easier because Distribution.__init__ is only
called from DistributionSet.new.