Remove deprecated ensure_unicode function
This is equivalent to `six.ensure_text` (strictly with
`encoding='ASCII'`, although that doesn't seem necessary in search
methods) if the input string is not None.  It may be that we could get
away without some of these `six.ensure_text` calls, but let's just keep
it simple for now; we can drop them more easily once we're on Python 3
everywhere.