testopenid: Use a filesystem-based store
Now that we normally use gunicorn for development setups, there are
multiple worker processes serving requests, and they won't have access
to the same MemoryStore instance for testopenid requests; this can cause
OpenID exchanges to fail depending on which workers serve the relevant
requests.  Use a filesystem-based store instead so that live nonces and
associations can be shared between processes.