Replace Loggerhead cookie handling using secure-cookie
paste.auth.cookie is deprecated, and getting it to work with Python 3 is
cumbersome at best.  Switch to secure-cookie, which seems to have a
nicer API for our purposes, appears to be better-maintained (by the
Werkzeug folks), and allows us to fix a long-standing minor deficiency
around deleting session cookies on logout.

This will break existing Loggerhead sessions.  I think that's OK since
it's a one-time thing, but it's worth noting.