Fix imprecise tests in xx-upstream-bug-privacy.txt
One of the examples in xx-upstream-bug-privacy.txt listed the bug number
and importance in the wrong order, but the test was so imprecise that it
normally managed to pass anyway by matching random bits of text from
things like HTML comments and the `LP.cache` JavaScript object, except
when dict ordering changes in Python 3 happened to put those in a
different order.  Make these tests more precise so that they don't
accidentally depend on this kind of thing.