Emit some more build metrics
We now increment a build.finished counter with the new build status name
when a build reaches a final state, and build.reset when it is reset.
In addition, number-cruncher sends a builders.failure_count gauge with
the current failure count of each builder.  This should be enough to
monitor some operational problems that result in high failure rates of
various kinds.