Send Content-Length when uploading OCI blobs to a registry
In Python 3.5, `tarfile.ExFileObject` has a `fileno` method that raises
`AttributeError`, which is enough to confuse `requests` into thinking
that it needs to do a chunked upload, which in turn causes `urllib3` not
to use the configured proxy.  Explicitly passing a Content-Length header
is polite anyway, and works around this misbehaviour.