Add a "deceased" account status
Sometimes Launchpad users die, and we're currently limited in what we
can do to record that information sensitively.  We generally want to at
least prevent people trying to contact them via Launchpad, and prevent
people who may have obtained the deceased person's credentials from
logging in.  If the user's family indicate that they want to hide or
remove the user's data, then existing mechanisms (deactivation or
close-account) are good enough, but otherwise it would be useful to have
a way to archive their account.

This adds a new `DECEASED` account status.  Accounts in that status will
have an appropriate message displayed at the top of their profile page,
and will be considered inactive so that they will not be able to log in.

In memory of Adam Conrad.