Add a WSGI authenticator for private PPAs
I got sufficiently annoyed in the process of fixing bug 1722209 to see
how hard it would be to fix it properly, since the existing htpasswd
scheme has been a thorn in our side for a while now.  The answer appears
to be "a bit, but not very".

This will let us remove our reliance on htpasswd files, and probably
eventually make it easier to do things like issue time-limited tokens
(or even macaroons?) to builders.

There are obviously various deployment issues to sort out, and
generate-ppa-htaccess still deals with things like deactivation and
cancellation emails which we'll need to move elsewhere; it may be
possible to just inline the relevant checks into the DB queries in
ArchiveAuthTokenSet, since we're always looking at a single subscriber
or named token name and so the result set sizes are very small.

The way that mod_wsgi loads the entry point is peculiar.  I considered
separating the client code off entirely from Launchpad, but this turns
out to be light enough in terms of memory use and startup time that I
don't think it's worth the effort at the moment.