Rename "Resubmit" vote to "Needs Resubmitting"
Merge proposal submitters often interpret the "Resubmit" vote to mean "I
have made the requested changes to this merge proposal and would like to
resubmit it for review", whereas in fact it's intended to be cast by a
reviewer and means "the idea might be sound but the implementation needs
significant rework".  Rename it so that it's clearer that it goes with
the "Needs Fixing" and "Needs Information" votes.