Clean up InProcessKeyServerFixture asynchronously
Tests that use `InProcessKeyServerFixture` fail anomalously often on
buildbot, and it may be because the fixture's cleanup isn't being done
quite right: `port.stopListening` may return a `Deferred`, and nothing
was waiting for that to fire.  Rearrange the cleanup code to run
asynchronously.  This is rather awkward in places due to the lack of a
proper asynchronous fixtures API.