Don't overwrite existing files in createEmptyPocketRequests
The job of `FTPArchive.createEmptyPocketRequests` is to make sure that
override and file list files for suites/components that we're publishing
exist even if they're going to end up empty.  It did this by making them
empty, which was OK when nothing was reading from those files in
parallel.  However, that hasn't been true for some time due to Contents
file generation, and the current implementation means that it's possible
for Contents file generation to see temporarily-empty versions of these
files and thus generate empty Contents files.

To fix this, only create these files if they don't already exist.  This
should complete the fix in
https://code.launchpad.net/~twom/launchpad/+git/launchpad/+merge/374930.

LP: #1871920