Generate empty override and file list files where necessary
Commit 25ce0c1d87 stopped `FTPArchive.createEmptyPocketRequests` from
writing empty override and file list files for all suites/components,
instead merely making sure that they exist.  However, this broke in the
case where a suite/component (or suite/component/architecture) was
non-empty and then became empty, because the previous non-empty
overrides and file lists were left in place.

To avoid this, ensure that `FTPArchive.publishOverrides` writes override
files for all relevant components, and that
`FTPArchive.publishFileLists` writes file list files for all relevant
components/architectures; these may be empty if there are no
publications in the corresponding contexts.  This doesn't suffer from
the problem that the aforementioned commit was fixing, because these
files are put in place atomically.

It's possible that this comes close to making
`FTPArchive.createEmptyPocketRequests` entirely unnecessary, but I
haven't yet tried to remove it.