Allow verifying SnapBuild/BPB macaroons for archives
As part of replacing `Archive.buildd_secret`, it'll be useful to have
tokens that can be used to access an archive but only for the lifetime
of a relevant build.  Since these tokens won't be owned by users and so
don't need to support manual revocation, it should be safe to reuse our
existing macaroon infrastructure for this application.

To start with, extend the binary-package-build and snap-build macaroon
issuers to support verification using an archive as the context.  This
currently requires the context archive to be either the build's archive
or one of its archive dependencies.

Nothing actually verifies these tokens yet, so there should be no
functional change as a result of this commit.

LP: #600910