Rebuild virtualenv if $(PYTHON) changes
On production, we build the tree on a central system, rsync it out, and
then run "make compile" on each target (or "make build" in the case of
frontends that need to ship CSS/JS).  This speeds up deployments: the
target systems *can* build the tree from scratch if need be, but they
normally don't have to.  However, when deploying some targets on Python
2 and some on Python 3, this doesn't quite work, because "make compile"
fails to rebuild the virtualenv on systems where we're asking for
`PYTHON=python3`.

To fix this, rebuild the virtualenv if `$(PYTHON)` doesn't match the
version of Python that was used to build the current virtualenv.