Remove env/ directory in "make clean_pip"
This target is used as part of the wheel-building machinery for
production deployments.  If we leave the env/ directory in place after
building wheels, then pushing code can get confused: since env/ still
exists, it thinks it doesn't need to rebuild it, but bin/ has been
removed so the tree won't work until it's been rebuilt.