Export a simple read-only API for polls
This is just enough to allow us to deal with poll spam more easily.  The
internal interface for selecting the ordering of poll results changes a
bit, but existing web UI views should remain unchanged.

We don't need to consider privacy, as the validator on `Poll.team_id`
already constrains polls to be owned only by public teams.