Pass behaviour rather than build to get_sources_list_for_building
It's straightforward to get a build from a build behaviour, but not
necessarily the other way round (since the behaviour has additional
state such as the builder).  Refactor get_sources_list_for_building and
its callers to pass down the behaviour, since that will be useful
shortly.

No functional change: this is just preparation for later work.