Dispatch sources.list entries for private archives using macaroons
This allows snap base archive dependencies on private archives to work
(for 16.04 ESM), and also takes a significant step towards abolishing
`Archive.buildd_secret`.  These tokens are only valid while the build is
running, and so are much less vulnerable to accidental leakage.

The old buildd secret is still used in the file map passed to builders
to fetch the source package for binary package builds in private
archives, so it can't yet be removed entirely.

LP: #600910