Add script to sync branches from production to staging
We currently have a "codehosting_branch_mirror.py" script on (qa)staging
that rsyncs a few branches from production for testing.  This is a
Python script that uses Launchpad's virtualenv, imports our modules, and
talks to our database.  As a result, it can easily be broken by changes
in Launchpad: in particular, it was broken by moving to Python 3.

Bring a version of this script into our tree so that we can make sure it
stays working.  I modified it considerably, mainly to allow selecting
branches on the command line, and added tests.