Fix creating /pillar/+oci/name via git push
This is a bit more involved than I'd initially anticipated, because OCI
projects don't have owners so there isn't an obvious natural choice for
the owner of the new repository.  For the time being, use the OCI
project admin in the case of distribution-based OCI projects, and
otherwise the pillar's owner.  I expect this may need to evolve further
in future.