Show hover text more consistently on datetimes
Many places in Launchpad's UI showed approximate/display dates with the
exact timestamp in a tooltip, but this was quite inconsistent.  Now that
we have some additional formatting helpers to make this easier, use them
in all the places where we previously showed only the
approximate/display date and where the underlying datetime refers to an
exact timestamp rather than an estimate.

A few tests required small adjustments to query counts, since the
tooltips show timestamps in the user's local time if possible, and that
requires a `PersonLocation` query.