Move remaining generate-ppa-htaccess functions to garbo
Now that archive authentication is handled dynamically, we no longer
need to deal with expiring archive subscriptions and deactivating
archive authorization tokens via the relatively heavyweight mechanism of
a standalone script run once per minute on the PPA publisher system:
it's really just a particular kind of garbage collection (with the only
quirk being that it sends emails), so it can reasonably be run from
garbo.

Deactivating tokens depends on expiring subscriptions, but the latter is
quite lightweight and so can be run from garbo-frequently, so in
practice this won't delay things for long, and it isn't vital for the
cancellation emails sent when tokens are deactivated to be sent
immediately.

We can remove generate-ppa-htaccess entirely once this commit lands on
production and once we've removed it from production crontabs.