Accept a comma-separated list in database.rw_main_slave
We currently have a rather complex arrangement of production config
files whose purpose is to distribute load between two standby databases
(although in fact both happen to point to the same actual database in
pgbouncer at the moment), and it would be useful to be able to simplify
this.

Accept a comma-separated list of connection strings, and select randomly
between them.  This should be good enough at distributing load if and
when we ever have multiple standby databases again, and will allow us to
disentangle the production configs a little bit.