Add basic model for charm recipe jobs
This job is a placeholder and doesn't actually do anything yet: we'll
need a model for charm recipe builds first.  I'm adding the job first
because I want to take the more modern approach of all charm recipe
builds having an associated build request.

We'll need to deploy a "charm-build-job" DB user to production before
landing this.