Avoid auto-repacking the same repository too often
We don't want to issue automatic repacks for the same Git repository too
often: if a repository keeps needing to be repacked on subsequent runs
of the cron job, it's much more likely to be because the repack queue
hasn't caught up yet than because it's accumulated much more data.
Throttle automatic repacks of any single repository to once a week.