librarianserver: Support multiple Swift instances
When migrating from one Swift instance to another (between clouds),
moving all the production data takes a long time.  Rather than having to
wait for this, it's better if we can start using the new Swift instance
in parallel with the data migration.  To support this, allow the
librarian to be configured with two sets of Swift credentials: in that
case, it will read from both (preferring the newer instance), feed new
files only into the newer instance, and garbage-collect from both
instances.