Refactor longhand versions of fmt:*datetitle
We now have `fmt:approximatedatetitle` and `fmt:displaydatetitle`, but
there were still quite a number of places that wrote out some rough
equivalent of those in longhand.  Use the new helpers for this.  The
HTML tag structure will change somewhat, as this means that we now use
the HTML5 `<time>` element, but the visual rendering should be as
before.