Move markSuiteDirty to IArchiveAppend
`markSuiteDirty` allows forcing a suite in an archive to be republished
as if by an upload without actually uploading to it.  It was previously
in `IArchiveEdit`, which requires the distribution owner in the case of
primary archives and the archive owner in all other cases.  This turns
out to be inconvenient in practice for Ubuntu, where the archive owner
is ~ubuntu-archive and handles day-to-day archive administration work,
and the distribution owner is ~techboard which doesn't.

There isn't an obvious perfect fit here, but `IArchiveAppend` seems a
lot closer: it requires the archive owner or somebody with upload
permission in the case of PPAs, and the archive owner in all other
cases.  The only significant downside is that Launchpad administrators
won't be able to use it any more, but I think this is tolerable.  If it
becomes a problem, we can probably press the `launchpad.Owner`
permission into service, since it's currently unused for archives.

LP: #1874859