setup.py: Filter out pip-build-env-* directories from system paths
I've occasionally seen pip-build-env-* directories leaking into
`sys.path` at the point when Launchpad's setup.py is run, which results
in picking up fragments of `sitecustomize.py` which then refer to
directories that no longer exist after `pip` exits, resulting in a
broken installation.  It's not completely clear to me why this happens,
but it's very confusing when it does, so filter those out.