Delete an incorrect tarfile_helpers doctest
This test only worked because it was passing a dict, which intrinsically
can't have duplicated keys; that says nothing about the behaviour of
`LaunchpadWriteTarFile`.  When I fixed that, the test failed because tar
files can in fact have multiple members with the same name!  I conclude
that this test isn't doing anything very useful and it's better to just
delete it.